On_close logic, first steps (read on).
The problem of SockJS till this point was that we
were unable to detect whether the response is still
open or whether it has been close by the client or
just by losing the connection. Whatever we do, we
just can't access sockets directly, hence we're
clueless about their state (thank YOU EventMachine
... ). So the best thing we can do is to write an
empty string to the response asking "Are you still
alive?" How very asynchronous indeed!

At this point, the tests don't pass yet, this is
the first step really, it's here more for the
commit message to explain what on Earth are we
doing.